requirements.txt
numpy
torch
torchvision
natsort
visdom
dominate
pyyaml
opencv-python
pillow
diffuser
xformers
matplotlib